requirements.txt
accelerate>=1.4.0
datasets>=3.2.0
huggingface_hub>=0.31.0
indic-transliteration>=2.3.59
librosa>=0.10.2
openai>=1.76.0
peft>=0.15.0
pydantic>=2.11.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
sentencepiece>=0.2.0
soundfile>=0.13.1
torch>=2.4.0
transformers>=4.57.0
pytest>=8.3.5
bitsandbytes>=0.45.0; platform_system != "Darwin"
